fix(plugin): guarantee plugin shutdown and terminal discards
Address review feedback on the bounded proposal dispatch path.

- Close queues instead of enqueueing a Shutdown command, so a full queue
  can no longer strand a worker and skip the native shutdown callback.
- Join a worker on an explicit exit signal rather than is_finished(), so
  the gap between the last callback and thread teardown cannot skip join.
- Classify a proposal's forwarding decision and telemetry from a single
  timestamp, so a candidate can never be forwarded while telemetry calls
  it late.
- Reserve queue headroom for terminal dispositions so a late candidate's
  discard is delivered instead of silently dropped.
- Surface a failing proposal callback as SourceError with the plugin's
  message instead of an unexplained silent abstention.
- Extract the dispatch subsystem into plugin_dispatch, taking lib.rs from
  1689 to 745 lines per the repo's >1000-line rule.
