+++ crates/mesh-native-serving-plugin-host/src/lib.rs
@@ -1,16 +1,15 @@
 //! Loads one native serving plugin and adapts its stable ABI to Skippy hooks.
 
+mod plugin_dispatch;
+#[cfg(test)]
+mod test_support;
+
 use std::{
-    collections::VecDeque,
     ffi::{c_char, c_void},
     path::{Path, PathBuf},
     ptr::NonNull,
-    sync::{
-        Arc, Condvar, Mutex, OnceLock,
-        atomic::{AtomicBool, AtomicU64, Ordering},
-        mpsc::{SyncSender, sync_channel},
-    },
-    thread::{self, JoinHandle},
+    sync::{Arc, Mutex, OnceLock},
+    thread,
     time::{Duration, Instant},
 };
 
@@ -22,17 +21,16 @@ use skippy_server::frontend::{
     GenerationReceipt, GenerationReceiptConfig, GenerationStart, LinearProposal,
     LinearProposalDiscardReason, LinearProposalDisposition, LinearProposalIngress,
     LinearProposalIngressConfig, LinearProposalQuery, LinearProposalReceipt,
-    LinearProposalSourceOutcome, LinearProposalSourceResponse, LinearProposalSourceTelemetry,
-    OpaqueProposalDecisionId,
+    LinearProposalSourceResponse, OpaqueProposalDecisionId,
 };
 use skippy_server::serving_hooks::{ModelServingHooks, ModelServingHooksFactory};
 use skippy_server::tokenizer::TokenizerCapability;
 
+use plugin_dispatch::{PluginCommand, PluginDriver};
+
 const ERROR_BUFFER_BYTES: usize = 2_048;
-const PLUGIN_COMMAND_CAPACITY: usize = 1_024;
 const MAX_NATIVE_PLUGIN_PROPOSAL_TOKENS: usize = 4_096;
 const PROPOSAL_POLL_INTERVAL: Duration = Duration::from_micros(50);
-const CLEAN_SHUTDOWN_TIMEOUT: Duration = Duration::from_secs(5);
 
 /// Mesh-owned factory for one independently built native serving plugin.
 #[derive(Clone)]
@@ -528,6 +526,10 @@ struct NativeProposalIngress {
 impl LinearProposalIngress for NativeProposalIngress {
     fn propose(&self, query: LinearProposalQuery) -> Result<LinearProposalSourceResponse> {
         let response = self.driver.propose(query)?;
+        // A failing plugin callback must not fail the generation, so a source
+        // error degrades to an abstention for decode. The worker logs the
+        // plugin's message and the `SourceError` outcome keeps the failure
+        // distinguishable from a deliberate abstention in telemetry.
         let proposal = response.proposal.unwrap_or_default();
         Ok(LinearProposalSourceResponse::with_telemetry(
             proposal,
@@ -548,428 +550,6 @@ impl LinearProposalIngress for NativeProposalIngress {
             decision_id.as_bytes().to_vec(),
             reason,
         ))
-    }
-}
-
-enum PluginCommand {
-    Begin(GenerationStart),
-    Committed(GenerationCommit),
-    Abort(GenerationAbort),
-    Finish(GenerationReceipt),
-    Proposal(LinearProposalQuery, SyncSender<ProposalResponse>),
-    Report(LinearProposalReceipt),
-    Discard(Vec<u8>, LinearProposalDiscardReason),
-    Shutdown(SyncSender<std::result::Result<(), String>>),
-}
-
-struct ProposalResponse {
-    proposal: std::result::Result<Option<LinearProposal>, String>,
-    telemetry: LinearProposalSourceTelemetry,
-}
-
-struct QueuedPluginCommand {
-    enqueued_at: Instant,
-    command: PluginCommand,
-}
-
-struct PluginCommandQueue {
-    commands: Mutex<VecDeque<QueuedPluginCommand>>,
-    stopped: AtomicBool,
-    available: Condvar,
-}
-
-#[derive(Debug)]
-enum PluginCommandQueueError {
-    Full,
-    Stopped,
-    Poisoned,
-}
-
-impl PluginCommandQueue {
-    fn new() -> Self {
-        Self {
-            commands: Mutex::new(VecDeque::with_capacity(PLUGIN_COMMAND_CAPACITY)),
-            stopped: AtomicBool::new(false),
-            available: Condvar::new(),
-        }
-    }
-
-    fn enqueue(&self, command: PluginCommand) -> Result<()> {
-        self.try_enqueue(command).map_err(|error| match error {
-            PluginCommandQueueError::Full => anyhow!("native serving plugin command queue is full"),
-            PluginCommandQueueError::Stopped => anyhow!("native serving plugin worker stopped"),
-            PluginCommandQueueError::Poisoned => {
-                anyhow!("native serving plugin command queue lock poisoned")
-            }
-        })
-    }
-
-    fn try_enqueue(
-        &self,
-        command: PluginCommand,
-    ) -> std::result::Result<(), PluginCommandQueueError> {
-        let mut commands = self
-            .commands
-            .lock()
-            .map_err(|_| PluginCommandQueueError::Poisoned)?;
-        if self.stopped.load(Ordering::Acquire) {
-            return Err(PluginCommandQueueError::Stopped);
-        }
-        if commands.len() == PLUGIN_COMMAND_CAPACITY {
-            return Err(PluginCommandQueueError::Full);
-        }
-        commands.push_back(QueuedPluginCommand {
-            enqueued_at: Instant::now(),
-            command,
-        });
-        self.available.notify_one();
-        Ok(())
-    }
-
-    fn mark_stopped(&self) {
-        self.stopped.store(true, Ordering::Release);
-        self.available.notify_all();
-    }
-
-    fn next(&self) -> QueuedPluginCommand {
-        let mut commands = self
-            .commands
-            .lock()
-            .expect("native serving plugin command queue lock must not be poisoned");
-        loop {
-            if let Some(command) = Self::pop_next(&mut commands) {
-                return command;
-            }
-            commands = self
-                .available
-                .wait(commands)
-                .expect("native serving plugin command queue lock must not be poisoned");
-        }
-    }
-
-    fn pop_next(commands: &mut VecDeque<QueuedPluginCommand>) -> Option<QueuedPluginCommand> {
-        // Lifecycle and proposal callbacks share this FIFO so every proposal
-        // observes its earlier committed state. Passive callbacks use their
-        // own worker and cannot delay either class.
-        commands.pop_front()
-    }
-}
-
-struct PluginDriver {
-    queue: Arc<PluginCommandQueue>,
-    passive_queue: Arc<PluginCommandQueue>,
-    active: Arc<ActivePlugin>,
-    fatal_error: Arc<Mutex<Option<String>>>,
-    lifecycle_delivery_failures: Arc<AtomicU64>,
-    worker: Mutex<Option<JoinHandle<()>>>,
-    passive_worker: Mutex<Option<JoinHandle<()>>>,
-}
-
-impl PluginDriver {
-    fn spawn(active: ActivePlugin) -> Result<Self> {
-        let queue = Arc::new(PluginCommandQueue::new());
-        let passive_queue = Arc::new(PluginCommandQueue::new());
-        let active = Arc::new(active);
-        let fatal_error = Arc::new(Mutex::new(None));
-        let worker_fatal_error = Arc::clone(&fatal_error);
-        let lifecycle_delivery_failures = Arc::new(AtomicU64::new(0));
-        let worker_lifecycle_delivery_failures = Arc::clone(&lifecycle_delivery_failures);
-        let worker_queue = Arc::clone(&queue);
-        let worker_passive_queue = Arc::clone(&passive_queue);
-        let worker_active = Arc::clone(&active);
-        let worker = thread::Builder::new()
-            .name("mesh-native-serving-plugin".to_string())
-            .spawn(move || {
-                plugin_worker(
-                    worker_active,
-                    worker_queue,
-                    worker_passive_queue,
-                    worker_fatal_error,
-                    worker_lifecycle_delivery_failures,
-                );
-            })
-            .context("spawn native serving plugin worker")?;
-        let passive_worker_queue = Arc::clone(&passive_queue);
-        let passive_worker_active = Arc::clone(&active);
-        let passive_worker = thread::Builder::new()
-            .name("mesh-native-serving-plugin-passive".to_string())
-            .spawn(move || plugin_passive_worker(passive_worker_active, passive_worker_queue))
-            .context("spawn native serving plugin passive worker")?;
-        Ok(Self {
-            queue,
-            passive_queue,
-            active,
-            fatal_error,
-            lifecycle_delivery_failures,
-            worker: Mutex::new(Some(worker)),
-            passive_worker: Mutex::new(Some(passive_worker)),
-        })
-    }
-
-    fn ensure_healthy(&self) -> Result<()> {
-        let error = self
-            .fatal_error
-            .lock()
-            .map_err(|_| anyhow!("native serving plugin health lock poisoned"))?;
-        if let Some(error) = error.as_deref() {
-            bail!("native serving plugin worker failed: {error}");
-        }
-        Ok(())
-    }
-
-    fn enqueue(&self, command: PluginCommand) -> Result<()> {
-        self.ensure_healthy()?;
-        self.enqueue_recovery(command)
-    }
-
-    /// Deliver lifecycle cleanup even after an earlier callback failed.
-    ///
-    /// This bypasses only the health gate. The command still uses its bounded
-    /// callback queue and fails if that worker has stopped or is full.
-    fn enqueue_recovery(&self, command: PluginCommand) -> Result<()> {
-        self.queue_for(&command).enqueue(command)
-    }
-
-    fn queue_for(&self, command: &PluginCommand) -> &Arc<PluginCommandQueue> {
-        if matches!(
-            command,
-            PluginCommand::Report(_) | PluginCommand::Discard(_, _)
-        ) {
-            &self.passive_queue
-        } else {
-            &self.queue
-        }
-    }
-
-    fn lifecycle_delivery_failures(&self) -> u64 {
-        self.lifecycle_delivery_failures.load(Ordering::Relaxed)
-    }
-
-    fn propose(&self, query: LinearProposalQuery) -> Result<ProposalResponse> {
-        self.ensure_healthy()?;
-        let deadline = query.deadline;
-        let submitted_at = Instant::now();
-        if submitted_at >= deadline {
-            return Ok(ProposalResponse {
-                proposal: Ok(None),
-                telemetry: LinearProposalSourceTelemetry {
-                    queue_wait_us: 0,
-                    callback_elapsed_us: 0,
-                    outcome: LinearProposalSourceOutcome::HostDeadlineExceeded,
-                },
-            });
-        }
-        let (reply, response) = sync_channel(1);
-        match self
-            .queue
-            .try_enqueue(PluginCommand::Proposal(query, reply))
-        {
-            Ok(()) => {}
-            Err(PluginCommandQueueError::Full) => {
-                return Ok(ProposalResponse {
-                    proposal: Ok(None),
-                    telemetry: LinearProposalSourceTelemetry {
-                        queue_wait_us: 0,
-                        callback_elapsed_us: 0,
-                        outcome: LinearProposalSourceOutcome::QueueFull,
-                    },
-                });
-            }
-            Err(PluginCommandQueueError::Stopped) => {
-                bail!("native serving plugin worker stopped before accepting proposal")
-            }
-            Err(PluginCommandQueueError::Poisoned) => {
-                bail!("native serving plugin command queue lock poisoned")
-            }
-        }
-        let remaining = deadline.saturating_duration_since(Instant::now());
-        if remaining.is_zero() {
-            return Ok(ProposalResponse {
-                proposal: Ok(None),
-                telemetry: LinearProposalSourceTelemetry {
-                    queue_wait_us: elapsed_us(submitted_at),
-                    callback_elapsed_us: 0,
-                    outcome: LinearProposalSourceOutcome::HostDeadlineExceeded,
-                },
-            });
-        }
-        match response.recv_timeout(remaining) {
-            Ok(result) => Ok(result),
-            Err(std::sync::mpsc::RecvTimeoutError::Timeout) => Ok(ProposalResponse {
-                proposal: Ok(None),
-                telemetry: LinearProposalSourceTelemetry {
-                    queue_wait_us: elapsed_us(submitted_at),
-                    callback_elapsed_us: 0,
-                    outcome: LinearProposalSourceOutcome::HostDeadlineExceeded,
-                },
-            }),
-            Err(std::sync::mpsc::RecvTimeoutError::Disconnected) => {
-                bail!("native serving plugin worker stopped before replying")
-            }
-        }
-    }
-}
-
-fn elapsed_us(started: Instant) -> u64 {
-    u64::try_from(started.elapsed().as_micros()).unwrap_or(u64::MAX)
-}
-
-impl Drop for PluginDriver {
-    fn drop(&mut self) {
-        self.stop_worker(&self.queue, &self.worker);
-        self.stop_worker(&self.passive_queue, &self.passive_worker);
-        if let Some(active) = Arc::get_mut(&mut self.active)
-            && let Err(error) = active.shutdown()
-        {
-            eprintln!("native serving plugin shutdown failed: {error:#}");
-        }
-    }
-}
-
-impl PluginDriver {
-    fn stop_worker(&self, queue: &Arc<PluginCommandQueue>, worker: &Mutex<Option<JoinHandle<()>>>) {
-        let (reply, response) = sync_channel(1);
-        let clean_shutdown = queue.enqueue(PluginCommand::Shutdown(reply)).is_ok()
-            && response.recv_timeout(CLEAN_SHUTDOWN_TIMEOUT).is_ok();
-        if (clean_shutdown || queue.stopped.load(Ordering::Acquire))
-            && let Ok(mut worker) = worker.lock()
-            && worker.as_ref().is_some_and(JoinHandle::is_finished)
-            && let Some(worker) = worker.take()
-        {
-            let _ = worker.join();
-        }
-    }
-}
-
-struct WorkerStopGuard {
-    queue: Arc<PluginCommandQueue>,
-}
-
-impl Drop for WorkerStopGuard {
-    fn drop(&mut self) {
-        self.queue.mark_stopped();
-    }
-}
-
-fn plugin_worker(
-    active: Arc<ActivePlugin>,
-    queue: Arc<PluginCommandQueue>,
-    passive_queue: Arc<PluginCommandQueue>,
-    fatal_error: Arc<Mutex<Option<String>>>,
-    lifecycle_delivery_failures: Arc<AtomicU64>,
-) {
-    let _stop_guard = WorkerStopGuard {
-        queue: Arc::clone(&queue),
-    };
-    loop {
-        let QueuedPluginCommand {
-            enqueued_at,
-            command,
-        } = queue.next();
-        let (result, terminal, lifecycle) = match command {
-            PluginCommand::Begin(event) => (active.begin(&event), false, true),
-            PluginCommand::Committed(event) => (active.committed(&event), false, true),
-            PluginCommand::Abort(event) => (active.abort(&event), false, true),
-            PluginCommand::Finish(event) => (active.finish(&event), false, true),
-            PluginCommand::Proposal(query, reply) => {
-                let queue_wait_us = elapsed_us(enqueued_at);
-                if Instant::now() >= query.deadline {
-                    let _ = reply.send(ProposalResponse {
-                        proposal: Ok(None),
-                        telemetry: LinearProposalSourceTelemetry {
-                            queue_wait_us,
-                            callback_elapsed_us: 0,
-                            outcome: LinearProposalSourceOutcome::DeadlineExceededBeforeDispatch,
-                        },
-                    });
-                    continue;
-                }
-                let callback_started = Instant::now();
-                let result = active.propose(query);
-                let callback_elapsed_us = elapsed_us(callback_started);
-                let candidate_was_late =
-                    matches!(&result, Ok(Some(_))) && Instant::now() >= query.deadline;
-                let outcome = match &result {
-                    Ok(Some(_)) if Instant::now() >= query.deadline => {
-                        LinearProposalSourceOutcome::CandidateReturnedTooLate
-                    }
-                    Ok(Some(_)) => LinearProposalSourceOutcome::Ready,
-                    Ok(None) if Instant::now() >= query.deadline => {
-                        LinearProposalSourceOutcome::DeadlineExceededInPlugin
-                    }
-                    Ok(None) => LinearProposalSourceOutcome::Abstained,
-                    Err(_) => LinearProposalSourceOutcome::SourceError,
-                };
-                if candidate_was_late && let Ok(Some(proposal)) = &result {
-                    let _ = passive_queue.enqueue(PluginCommand::Discard(
-                        proposal.decision_id.as_bytes().to_vec(),
-                        LinearProposalDiscardReason::DeadlineExceeded,
-                    ));
-                }
-                let _ = reply.send(ProposalResponse {
-                    proposal: if candidate_was_late {
-                        Ok(None)
-                    } else {
-                        result
-                            .as_ref()
-                            .map(Clone::clone)
-                            .map_err(ToString::to_string)
-                    },
-                    telemetry: LinearProposalSourceTelemetry {
-                        queue_wait_us,
-                        callback_elapsed_us,
-                        outcome,
-                    },
-                });
-                (result.map(|_| ()), false, false)
-            }
-            PluginCommand::Shutdown(reply) => {
-                let _ = reply.send(Ok(()));
-                (Ok(()), true, false)
-            }
-            PluginCommand::Report(_) | PluginCommand::Discard(_, _) => {
-                unreachable!("passive plugin callbacks must use the passive worker queue")
-            }
-        };
-        if lifecycle && result.is_err() {
-            lifecycle_delivery_failures.fetch_add(1, Ordering::Relaxed);
-        }
-        if terminal
-            && let Err(error) = result
-            && let Ok(mut fatal) = fatal_error.lock()
-        {
-            *fatal = Some(format!("{error:#}"));
-        }
-        if terminal {
-            break;
-        }
-    }
-}
-
-fn plugin_passive_worker(active: Arc<ActivePlugin>, queue: Arc<PluginCommandQueue>) {
-    let _stop_guard = WorkerStopGuard {
-        queue: Arc::clone(&queue),
-    };
-    loop {
-        match queue.next().command {
-            PluginCommand::Report(event) => {
-                let _ = active.report(&event);
-            }
-            PluginCommand::Discard(decision_id, reason) => {
-                let _ = active.discard(&decision_id, reason);
-            }
-            PluginCommand::Shutdown(reply) => {
-                let _ = reply.send(Ok(()));
-                break;
-            }
-            PluginCommand::Begin(_)
-            | PluginCommand::Committed(_)
-            | PluginCommand::Abort(_)
-            | PluginCommand::Finish(_)
-            | PluginCommand::Proposal(_, _) => {
-                unreachable!("lifecycle and proposal callbacks must use the primary worker queue")
-            }
-        }
     }
 }
 
@@ -1083,10 +663,7 @@ unsafe fn read_utf8(slice: abi::ByteSlice, label: &str) -> Result<String> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::sync::atomic::{AtomicUsize, Ordering};
-
-    static FAKE_NAME: &[u8] = b"test-serving-plugin";
-    static CANCEL_COUNT: AtomicUsize = AtomicUsize::new(0);
+    use crate::test_support::fake_table;
 
     #[test]
     fn tokenizer_inventory_view_borrows_host_owned_bytes_for_activation() {
@@ -1112,251 +689,6 @@ mod tests {
             unsafe { std::slice::from_raw_parts(entry.bytes.pointer, entry.bytes.length) },
             b"hello"
         );
-    }
-
-    struct FakeState {
-        start_delay: Duration,
-        poll_delay: Duration,
-        poll_returns_candidate: bool,
-        commit_delay: Duration,
-        report_delay: Duration,
-        begin_fails: bool,
-        events: Arc<Mutex<Vec<&'static str>>>,
-        abort_count: Arc<AtomicUsize>,
-    }
-
-    unsafe extern "C" fn fake_activate(
-        _context: *const abi::ActivationContext,
-        _activation: *mut abi::PluginActivation,
-    ) -> abi::PluginStatus {
-        abi::PluginStatus::INTERNAL_ERROR
-    }
-
-    unsafe extern "C" fn fake_shutdown(instance: abi::PluginInstance) -> abi::PluginStatus {
-        if !instance.is_null() {
-            drop(unsafe { Box::from_raw(instance.cast::<FakeState>()) });
-        }
-        abi::PluginStatus::OK
-    }
-
-    unsafe extern "C" fn fake_begin(
-        instance: abi::PluginInstance,
-        _event: *const abi::GenerationStart,
-    ) -> abi::PluginStatus {
-        let state = unsafe { &*instance.cast::<FakeState>() };
-        state.events.lock().unwrap().push("begin");
-        if state.begin_fails {
-            abi::PluginStatus::INTERNAL_ERROR
-        } else {
-            abi::PluginStatus::OK
-        }
-    }
-
-    unsafe extern "C" fn fake_commit(
-        instance: abi::PluginInstance,
-        _event: *const abi::GenerationCommit,
-    ) -> abi::PluginStatus {
-        let state = unsafe { &*instance.cast::<FakeState>() };
-        state.events.lock().unwrap().push("commit");
-        thread::sleep(state.commit_delay);
-        abi::PluginStatus::OK
-    }
-
-    unsafe extern "C" fn fake_abort(
-        instance: abi::PluginInstance,
-        _event: *const abi::GenerationAbort,
-    ) -> abi::PluginStatus {
-        let state = unsafe { &*instance.cast::<FakeState>() };
-        state.abort_count.fetch_add(1, Ordering::SeqCst);
-        abi::PluginStatus::OK
-    }
-
-    unsafe extern "C" fn fake_finish(
-        _instance: abi::PluginInstance,
-        _event: *const abi::GenerationFinish,
-    ) -> abi::PluginStatus {
-        abi::PluginStatus::OK
-    }
-
-    unsafe extern "C" fn fake_start_proposal(
-        instance: abi::PluginInstance,
-        _query: *const abi::ProposalQuery,
-        operation: *mut abi::ProposalOperation,
-    ) -> abi::PluginStatus {
-        let state = unsafe { &*instance.cast::<FakeState>() };
-        state.events.lock().unwrap().push("proposal");
-        thread::sleep(state.start_delay);
-        unsafe { *operation = 1 };
-        abi::PluginStatus::OK
-    }
-
-    unsafe extern "C" fn fake_poll_proposal(
-        instance: abi::PluginInstance,
-        _operation: abi::ProposalOperation,
-        output: *mut abi::ProposalOutput,
-    ) -> abi::ProposalPollStatus {
-        let state = unsafe { &*instance.cast::<FakeState>() };
-        thread::sleep(state.poll_delay);
-        if !state.poll_returns_candidate {
-            return abi::ProposalPollStatus::ABSTAIN;
-        }
-        let output = unsafe { &mut *output };
-        unsafe {
-            *output.decision_id = 7;
-            *output.token_ids = 42;
-        }
-        output.decision_id_length = 1;
-        output.token_length = 1;
-        abi::ProposalPollStatus::READY
-    }
-
-    unsafe extern "C" fn fake_cancel_proposal(
-        _instance: abi::PluginInstance,
-        _operation: abi::ProposalOperation,
-    ) {
-        CANCEL_COUNT.fetch_add(1, Ordering::SeqCst);
-    }
-
-    unsafe extern "C" fn fake_report_proposal(
-        instance: abi::PluginInstance,
-        _event: *const abi::ProposalOutcome,
-    ) -> abi::PluginStatus {
-        let state = unsafe { &*instance.cast::<FakeState>() };
-        state.events.lock().unwrap().push("report");
-        thread::sleep(state.report_delay);
-        abi::PluginStatus::OK
-    }
-
-    unsafe extern "C" fn fake_discard_proposal(
-        instance: abi::PluginInstance,
-        _event: *const abi::ProposalDiscard,
-    ) -> abi::PluginStatus {
-        let state = unsafe { &*instance.cast::<FakeState>() };
-        state.events.lock().unwrap().push("discard");
-        thread::sleep(state.report_delay);
-        abi::PluginStatus::OK
-    }
-
-    unsafe extern "C" fn fake_last_error(
-        _instance: abi::PluginInstance,
-        _output: *mut c_char,
-        _capacity: usize,
-    ) -> usize {
-        0
-    }
-
-    fn fake_table() -> abi::NativeServingPluginV1 {
-        abi::NativeServingPluginV1 {
-            abi_version: abi::NATIVE_SERVING_PLUGIN_ABI_V1,
-            struct_size: size_of::<abi::NativeServingPluginV1>(),
-            plugin_name: abi::ByteSlice::from_bytes(FAKE_NAME),
-            activate: fake_activate,
-            shutdown: fake_shutdown,
-            begin_generation: fake_begin,
-            commit_generation: fake_commit,
-            abort_generation: fake_abort,
-            finish_generation: fake_finish,
-            start_proposal: fake_start_proposal,
-            poll_proposal: fake_poll_proposal,
-            cancel_proposal: fake_cancel_proposal,
-            report_proposal: fake_report_proposal,
-            discard_proposal: fake_discard_proposal,
-            last_error: fake_last_error,
-        }
-    }
-
-    fn fake_active(start_delay: Duration) -> ActivePlugin {
-        fake_active_with_events(start_delay).0
-    }
-
-    fn fake_active_with_events(
-        start_delay: Duration,
-    ) -> (ActivePlugin, Arc<Mutex<Vec<&'static str>>>) {
-        let (active, events, _) = fake_active_with_observations(start_delay);
-        (active, events)
-    }
-
-    fn fake_active_with_observations(
-        start_delay: Duration,
-    ) -> (
-        ActivePlugin,
-        Arc<Mutex<Vec<&'static str>>>,
-        Arc<AtomicUsize>,
-    ) {
-        fake_active_with_options(start_delay, false)
-    }
-
-    fn fake_active_with_options(
-        start_delay: Duration,
-        begin_fails: bool,
-    ) -> (
-        ActivePlugin,
-        Arc<Mutex<Vec<&'static str>>>,
-        Arc<AtomicUsize>,
-    ) {
-        fake_active_with_timing(start_delay, Duration::ZERO, Duration::ZERO, begin_fails)
-    }
-
-    fn fake_active_with_timing(
-        start_delay: Duration,
-        commit_delay: Duration,
-        report_delay: Duration,
-        begin_fails: bool,
-    ) -> (
-        ActivePlugin,
-        Arc<Mutex<Vec<&'static str>>>,
-        Arc<AtomicUsize>,
-    ) {
-        let table = Box::leak(Box::new(fake_table()));
-        let definition = Arc::new(LoadedDefinition {
-            _library: None,
-            api: NonNull::from(table),
-            name: "test-serving-plugin".to_string(),
-        });
-        let events = Arc::new(Mutex::new(Vec::new()));
-        let abort_count = Arc::new(AtomicUsize::new(0));
-        let state = Box::new(FakeState {
-            start_delay,
-            poll_delay: Duration::ZERO,
-            poll_returns_candidate: false,
-            commit_delay,
-            report_delay,
-            begin_fails,
-            events: Arc::clone(&events),
-            abort_count: Arc::clone(&abort_count),
-        });
-        (
-            ActivePlugin {
-                definition,
-                instance: NonNull::new(Box::into_raw(state).cast::<c_void>()),
-                proposal_token_buffer: Mutex::new(vec![0; MAX_NATIVE_PLUGIN_PROPOSAL_TOKENS]),
-            },
-            events,
-            abort_count,
-        )
-    }
-
-    fn fake_active_with_late_candidate(poll_delay: Duration) -> ActivePlugin {
-        let (active, _, _) =
-            fake_active_with_timing(Duration::ZERO, Duration::ZERO, Duration::ZERO, false);
-        let instance = active.instance.unwrap().as_ptr().cast::<FakeState>();
-        unsafe {
-            (*instance).poll_delay = poll_delay;
-            (*instance).poll_returns_candidate = true;
-        }
-        active
-    }
-
-    fn wait_for_event(events: &Mutex<Vec<&'static str>>, event: &str) {
-        let deadline = Instant::now() + Duration::from_secs(1);
-        while !events.lock().unwrap().contains(&event) {
-            assert!(Instant::now() < deadline, "timed out waiting for {event}");
-            thread::yield_now();
-        }
-    }
-
-    fn proposal_query(deadline: Instant) -> LinearProposalQuery {
-        LinearProposalQuery::new(1, 2, 1, 1, 0, 8, deadline)
     }
 
     #[test]
@@ -1409,281 +741,5 @@ mod tests {
                 .to_string()
                 .contains("table size")
         );
-    }
-
-    #[test]
-    fn blocking_plugin_cannot_extend_the_decode_deadline() {
-        CANCEL_COUNT.store(0, Ordering::SeqCst);
-        let driver = PluginDriver::spawn(fake_active(Duration::from_millis(250))).unwrap();
-        let started = Instant::now();
-        let result = driver
-            .propose(LinearProposalQuery::new(
-                1,
-                2,
-                16,
-                16,
-                0,
-                8_192,
-                started + Duration::from_millis(5),
-            ))
-            .unwrap();
-        let elapsed = started.elapsed();
-
-        assert!(result.proposal.unwrap().is_none());
-        assert_eq!(
-            result.telemetry.outcome,
-            LinearProposalSourceOutcome::HostDeadlineExceeded
-        );
-        assert!(
-            elapsed < Duration::from_millis(150),
-            "decode waited {elapsed:?} for a blocking plugin"
-        );
-        drop(driver);
-        assert_eq!(CANCEL_COUNT.load(Ordering::SeqCst), 1);
-    }
-
-    #[test]
-    fn slow_commit_abstains_before_plugin_dispatch_and_later_positions_recover() {
-        let (active, events, _) = fake_active_with_timing(
-            Duration::ZERO,
-            Duration::from_millis(40),
-            Duration::ZERO,
-            false,
-        );
-        let driver = Arc::new(PluginDriver::spawn(active).unwrap());
-        let ingress = NativeLifecycleIngress {
-            driver: Arc::clone(&driver),
-        };
-        ingress
-            .try_submit(GenerationLifecycleObservation::Committed(
-                GenerationCommit {
-                    request_id: 1,
-                    session_id: 2,
-                    generated_token_count: 1,
-                    token_ids: vec![4].into_boxed_slice(),
-                },
-            ))
-            .unwrap();
-        wait_for_event(events.as_ref(), "commit");
-
-        let started = Instant::now();
-        let missed = driver
-            .propose(proposal_query(started + Duration::from_millis(5)))
-            .unwrap();
-        assert!(missed.proposal.unwrap().is_none());
-        assert_eq!(
-            missed.telemetry.outcome,
-            LinearProposalSourceOutcome::HostDeadlineExceeded
-        );
-        assert!(
-            started.elapsed() < Duration::from_millis(30),
-            "proposal wait exceeded its deadline: {:?}",
-            started.elapsed()
-        );
-
-        let recovered = driver
-            .propose(proposal_query(Instant::now() + Duration::from_millis(100)))
-            .unwrap();
-        assert!(recovered.proposal.unwrap().is_none());
-        assert_eq!(
-            recovered.telemetry.outcome,
-            LinearProposalSourceOutcome::Abstained
-        );
-        assert_eq!(*events.lock().unwrap(), ["commit", "proposal"]);
-    }
-
-    #[test]
-    fn running_passive_discard_cannot_delay_the_next_proposal() {
-        let (active, events, _) = fake_active_with_timing(
-            Duration::ZERO,
-            Duration::ZERO,
-            Duration::from_millis(100),
-            false,
-        );
-        let driver = Arc::new(PluginDriver::spawn(active).unwrap());
-        driver
-            .enqueue(PluginCommand::Discard(
-                vec![1],
-                LinearProposalDiscardReason::PositionMismatch,
-            ))
-            .unwrap();
-        wait_for_event(events.as_ref(), "discard");
-
-        let started = Instant::now();
-        let response = driver
-            .propose(proposal_query(started + Duration::from_millis(20)))
-            .unwrap();
-        assert!(response.proposal.unwrap().is_none());
-        assert_eq!(
-            response.telemetry.outcome,
-            LinearProposalSourceOutcome::Abstained
-        );
-        assert!(started.elapsed() < Duration::from_millis(60));
-        assert_eq!(*events.lock().unwrap(), ["discard", "proposal"]);
-    }
-
-    #[test]
-    fn worker_reports_pre_dispatch_deadlines_without_running_the_callback() {
-        let (active, events, _) = fake_active_with_timing(
-            Duration::ZERO,
-            Duration::from_millis(40),
-            Duration::ZERO,
-            false,
-        );
-        let driver = PluginDriver::spawn(active).unwrap();
-        driver
-            .enqueue(PluginCommand::Committed(GenerationCommit {
-                request_id: 1,
-                session_id: 2,
-                generated_token_count: 1,
-                token_ids: vec![4].into_boxed_slice(),
-            }))
-            .unwrap();
-        wait_for_event(events.as_ref(), "commit");
-        let (reply, response) = sync_channel(1);
-        driver
-            .queue
-            .try_enqueue(PluginCommand::Proposal(
-                proposal_query(Instant::now() + Duration::from_millis(5)),
-                reply,
-            ))
-            .unwrap();
-
-        let response = response.recv_timeout(Duration::from_millis(100)).unwrap();
-        assert!(response.proposal.unwrap().is_none());
-        assert_eq!(
-            response.telemetry.outcome,
-            LinearProposalSourceOutcome::DeadlineExceededBeforeDispatch
-        );
-        assert_eq!(*events.lock().unwrap(), ["commit"]);
-    }
-
-    #[test]
-    fn late_candidate_is_reported_and_not_forwarded_to_the_decode() {
-        let driver =
-            PluginDriver::spawn(fake_active_with_late_candidate(Duration::from_millis(20)))
-                .unwrap();
-        let (reply, response) = sync_channel(1);
-        driver
-            .queue
-            .try_enqueue(PluginCommand::Proposal(
-                proposal_query(Instant::now() + Duration::from_millis(5)),
-                reply,
-            ))
-            .unwrap();
-
-        let response = response.recv_timeout(Duration::from_millis(100)).unwrap();
-        assert!(response.proposal.unwrap().is_none());
-        assert_eq!(
-            response.telemetry.outcome,
-            LinearProposalSourceOutcome::CandidateReturnedTooLate
-        );
-    }
-
-    #[test]
-    fn stopped_worker_rejects_lifecycle_delivery() {
-        let queue = PluginCommandQueue::new();
-        queue.mark_stopped();
-
-        assert!(matches!(
-            queue.try_enqueue(PluginCommand::Abort(GenerationAbort {
-                request_id: 1,
-                session_id: 2,
-            })),
-            Err(PluginCommandQueueError::Stopped)
-        ));
-    }
-
-    #[test]
-    fn lifecycle_ingress_shares_plugin_queue_order_with_proposals() {
-        let (active, events) = fake_active_with_events(Duration::ZERO);
-        let driver = Arc::new(PluginDriver::spawn(active).unwrap());
-        let ingress = NativeLifecycleIngress {
-            driver: Arc::clone(&driver),
-        };
-        ingress
-            .try_submit(GenerationLifecycleObservation::Started(GenerationStart {
-                request_id: 1,
-                session_id: 2,
-                agent_session_id: None,
-                prompt_token_ids: Arc::from([3]),
-            }))
-            .unwrap();
-        ingress
-            .try_submit(GenerationLifecycleObservation::Committed(
-                GenerationCommit {
-                    request_id: 1,
-                    session_id: 2,
-                    generated_token_count: 1,
-                    token_ids: vec![4].into_boxed_slice(),
-                },
-            ))
-            .unwrap();
-        driver
-            .propose(LinearProposalQuery::new(
-                1,
-                2,
-                1,
-                1,
-                0,
-                8,
-                Instant::now() + Duration::from_millis(100),
-            ))
-            .unwrap();
-
-        assert_eq!(*events.lock().unwrap(), ["begin", "commit", "proposal"]);
-    }
-
-    #[test]
-    fn lifecycle_callback_failure_is_observed_without_poisoning_the_driver() {
-        let (active, _, _) = fake_active_with_options(Duration::ZERO, true);
-        let driver = Arc::new(PluginDriver::spawn(active).unwrap());
-        let ingress = NativeLifecycleIngress {
-            driver: Arc::clone(&driver),
-        };
-        ingress
-            .try_submit(GenerationLifecycleObservation::Started(GenerationStart {
-                request_id: 7,
-                session_id: 9,
-                agent_session_id: None,
-                prompt_token_ids: Arc::from([3]),
-            }))
-            .unwrap();
-
-        driver
-            .propose(LinearProposalQuery::new(
-                7,
-                9,
-                1,
-                1,
-                0,
-                8,
-                Instant::now() + Duration::from_millis(100),
-            ))
-            .unwrap();
-
-        assert_eq!(driver.lifecycle_delivery_failures(), 1);
-        assert!(driver.ensure_healthy().is_ok());
-    }
-
-    #[test]
-    fn generation_abort_bypasses_unhealthy_driver_gate() {
-        let (active, _, abort_count) = fake_active_with_observations(Duration::ZERO);
-        let driver = Arc::new(PluginDriver::spawn(active).unwrap());
-        *driver.fatal_error.lock().unwrap() = Some("report proposal failed".to_string());
-        let ingress = NativeLifecycleIngress {
-            driver: Arc::clone(&driver),
-        };
-
-        ingress
-            .try_submit(GenerationLifecycleObservation::Aborted(GenerationAbort {
-                request_id: 7,
-                session_id: 9,
-            }))
-            .unwrap();
-
-        drop(ingress);
-        drop(driver);
-        assert_eq!(abort_count.load(Ordering::SeqCst), 1);
     }
 }
